backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

langchain-openai==0.1.8
langchain-core==0.2.5

redis==5.0.4

pymysql==1.1.1
sqlalchemy==2.0.30

cryptography==42.0.8